refactor: Split dashboard_routes/_routes.py into domain-specific modules
Extract 7 domain route groups from the monolithic _routes.py (4,178 lines)
into separate modules, reducing it to ~2,019 lines. Each domain file
exports a register(router, ctx) function called by create_router().

New modules:
- _epic_routes.py (3 routes)
- _crates_routes.py (9 routes)
- _hitl_routes.py (9 routes)
- _control_routes.py (13 routes, incl. admin + bot-pr)
- _metrics_routes.py (12 routes)
- _reports_routes.py (5 routes)
- _state_routes.py (19 routes, runtimes/repos/filesystem/github)

Mechanical refactor only — no logic changes. All 381 tests pass.
