+++ src/dashboard_routes/_routes.py
@@ -12,34 +12,28 @@ import re
 import sys
 import tempfile
 import time
-from collections import Counter
 from collections.abc import Awaitable, Callable, Iterable, Mapping
 from dataclasses import dataclass, field
-from datetime import UTC, datetime, timedelta
+from datetime import UTC, datetime
 from pathlib import Path, PurePosixPath
-from typing import TYPE_CHECKING, Any, Literal
+from typing import TYPE_CHECKING, Any
 
 from fastapi import (
     APIRouter,
-    Body,
     HTTPException,
-    Query,
     Response,
     WebSocket,
     WebSocketDisconnect,
 )
 from fastapi.responses import FileResponse, HTMLResponse, JSONResponse
-from pydantic import ValidationError
 
-from admin_tasks import TaskResult, run_clean, run_ensure_labels, run_prep, run_scaffold
+from admin_tasks import TaskResult
 from app_version import get_app_version
-from config import HydraFlowConfig, save_config_file
+from config import HydraFlowConfig
 from dashboard_routes._common import (
     _EPIC_INTERNAL_LABELS,
     _FRONTEND_STAGE_TO_LABEL_FIELD,
     _INFERENCE_COUNTER_KEYS,
-    _INTERVAL_BOUNDS,
-    _SAFE_SLUG_COMPONENT,
     _STAGE_NAME_MAP,
     _coerce_history_status,
     _coerce_int,
@@ -51,23 +45,10 @@ from dashboard_routes._common import (
 from events import EventBus, EventType, HydraFlowEvent
 from github_cache import GitHubDataCache
 from issue_fetcher import IssueFetcher
-from metrics_manager import get_metrics_cache_dir
 from models import (
-    BackgroundWorkersResponse,
-    BackgroundWorkerState,
-    BackgroundWorkerStatus,
     BGWorkerHealth,
-    ControlStatus,
-    ControlStatusConfig,
-    ControlStatusResponse,
-    CrateCreateRequest,
-    CrateItemsRequest,
-    CrateUpdateRequest,
     GitHubIssue,
-    HITLCloseRequest,
     HITLEscalationPayload,
-    HITLSkipRequest,
-    HITLUpdatePayload,
     IntentRequest,
     IntentResponse,
     IssueHistoryEntry,
@@ -76,21 +57,10 @@ from models import (
     IssueHistoryResponse,
     IssueOutcome,
     IssueOutcomeType,
-    MetricsHistoryResponse,
-    MetricsResponse,
-    MetricsSnapshot,
-    OrchestratorStatusPayload,
-    PendingReport,
     PipelineIssue,
     PipelineSnapshot,
     PipelineSnapshotEntry,
     QueueStats,
-    ReportHistoryEntry,
-    ReportIssueRequest,
-    ReportIssueResponse,
-    RepoRuntimeInfo,
-    TrackedReport,
-    TrackedReportUpdate,
     parse_task_links,
 )
 from pr_manager import PRManager
@@ -99,7 +69,6 @@ from route_types import RepoSlugParam
 from state import StateTracker
 from timeline import TimelineBuilder
 from transcript_summarizer import TranscriptSummarizer
-from update_check import load_cached_update_result
 
 if TYPE_CHECKING:
     from orchestrator import HydraFlowOrchestrator
@@ -709,38 +678,6 @@ def create_router(
         issue_number: int, *, cause: str, origin: str | None
     ) -> None:
         await ctx.warm_hitl_summary(issue_number, cause=cause, origin=origin)
-
-    def _load_local_metrics_cache(
-        target_config: HydraFlowConfig,
-        limit: int = 100,
-    ) -> list[MetricsSnapshot]:
-        """Load metrics snapshots from local disk cache without requiring the orchestrator."""
-        cache_file = get_metrics_cache_dir(target_config) / "snapshots.jsonl"
-        if not cache_file.exists():
-            return []
-        snapshots: list[MetricsSnapshot] = []
-        try:
-            with open(cache_file) as f:
-                for raw_line in f:
-                    stripped = raw_line.strip()
-                    if not stripped:
-                        continue
-                    try:
-                        snapshots.append(MetricsSnapshot.model_validate_json(stripped))
-                    except ValidationError:
-                        logger.debug(
-                            "Skipping corrupt metrics snapshot line",
-                            exc_info=True,
-                        )
-                        continue
-        except OSError:
-            logger.warning(
-                "Could not read metrics cache %s",
-                cache_file,
-                exc_info=True,
-            )
-            return []
-        return snapshots[-limit:]
 
     def _build_history_links(
         raw: dict[int, dict[str, Any]] | Iterable[Any],
@@ -1569,1084 +1506,30 @@ def create_router(
             [item if isinstance(item, dict) else item.model_dump() for item in items]
         )
 
-    @router.get("/api/epics")
-    async def get_epics(
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Return all tracked epics with enriched sub-issue progress."""
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        orch = _get_orch()
-        if orch is None:
-            return JSONResponse([])
-        details = await orch._svc.epic_manager.get_all_detail()
-        return JSONResponse([d.model_dump() for d in details])
+    # --- Epic routes (extracted to _epic_routes.py) ---
+    from dashboard_routes._epic_routes import register as _register_epics
 
-    @router.get("/api/epics/{epic_number}")
-    async def get_epic_detail(
-        epic_number: int,
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Return full detail for a single epic including child issue info."""
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        orch = _get_orch()
-        if orch is None:
-            return JSONResponse({"error": "orchestrator not running"}, status_code=503)
-        detail = await orch._svc.epic_manager.get_detail(epic_number)
-        if detail is None:
-            return JSONResponse({"error": "epic not found"}, status_code=404)
-        return JSONResponse(detail.model_dump())
+    _register_epics(router, ctx)
 
-    @router.post("/api/epics/{epic_number}/release")
-    async def trigger_epic_release(
-        epic_number: int,
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Trigger async merge sequence and release creation for an epic.
+    # --- Crate routes (extracted to _crates_routes.py) ---
+    from dashboard_routes._crates_routes import register as _register_crates
 
-        Returns a job_id. Completion is signalled via the EPIC_RELEASED WebSocket
-        event — there is no REST polling endpoint for job status.
-        """
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        orch = _get_orch()
-        if orch is None:
-            return JSONResponse({"error": "orchestrator not running"}, status_code=503)
-        result = await orch._svc.epic_manager.trigger_release(epic_number)
-        if "error" in result:
-            return JSONResponse(result, status_code=400)
-        return JSONResponse(result)
+    _register_crates(router, ctx)
 
-    # --- Crate (milestone) routes ---
+    # --- HITL routes (extracted to _hitl_routes.py) ---
+    from dashboard_routes._hitl_routes import register as _register_hitl
 
-    @router.get("/api/crates")
-    async def get_crates() -> JSONResponse:
-        """List all milestones as crates with enriched progress data."""
-        try:
-            crates = await pr_manager.list_milestones()
-            result = []
-            for crate in crates:
-                data = crate.model_dump()
-                data["total_issues"] = crate.open_issues + crate.closed_issues
-                data["progress"] = (
-                    round(
-                        crate.closed_issues
-                        / (crate.open_issues + crate.closed_issues)
-                        * 100
-                    )
-                    if (crate.open_issues + crate.closed_issues) > 0
-                    else 0
-                )
-                result.append(data)
-            return JSONResponse(result)
-        except RuntimeError as exc:
-            logger.error("Failed to fetch crates: %s", exc)
-            return JSONResponse({"error": "Failed to fetch crates"}, status_code=500)
+    _register_hitl(router, ctx)
 
-    @router.post("/api/crates")
-    async def create_crate(body: CrateCreateRequest) -> JSONResponse:
-        """Create a new milestone (crate)."""
-        if not body.title.strip():
-            return JSONResponse({"error": "title is required"}, status_code=400)
-        try:
-            crate = await pr_manager.create_milestone(
-                title=body.title.strip(),
-                description=body.description,
-                due_on=body.due_on,
-            )
-            return JSONResponse(crate.model_dump())
-        except RuntimeError as exc:
-            logger.error("Failed to create crate: %s", exc)
-            return JSONResponse({"error": "Failed to create crate"}, status_code=500)
+    # --- Control routes (extracted to _control_routes.py) ---
+    from dashboard_routes._control_routes import register as _register_control
 
-    @router.patch("/api/crates/{crate_number}")
-    async def update_crate(crate_number: int, body: CrateUpdateRequest) -> JSONResponse:
-        """Update a milestone (crate).
+    _register_control(router, ctx)
 
-        Only fields present in the request JSON are forwarded.  Sending
-        ``"due_on": null`` clears the milestone due date.
-        """
-        fields = {k: body.model_dump()[k] for k in body.model_fields_set}
-        if not fields:
-            return JSONResponse({"error": "no fields to update"}, status_code=400)
-        try:
-            crate = await pr_manager.update_milestone(crate_number, **fields)
-            return JSONResponse(crate.model_dump())
-        except RuntimeError as exc:
-            logger.error("Failed to update crate #%d: %s", crate_number, exc)
-            return JSONResponse({"error": "Failed to update crate"}, status_code=500)
+    # --- Metrics routes (extracted to _metrics_routes.py) ---
+    from dashboard_routes._metrics_routes import register as _register_metrics
 
-    @router.delete("/api/crates/{crate_number}")
-    async def delete_crate(crate_number: int) -> JSONResponse:
-        """Delete a milestone (crate)."""
-        try:
-            await pr_manager.delete_milestone(crate_number)
-            return JSONResponse({"ok": True})
-        except RuntimeError as exc:
-            logger.error("Failed to delete crate #%d: %s", crate_number, exc)
-            return JSONResponse({"error": "Failed to delete crate"}, status_code=500)
-
-    @router.post("/api/crates/{crate_number}/items")
-    async def add_crate_items(
-        crate_number: int, body: CrateItemsRequest
-    ) -> JSONResponse:
-        """Assign issues to a milestone (crate)."""
-        try:
-            for issue_number in body.issue_numbers:
-                await pr_manager.set_issue_milestone(issue_number, crate_number)
-            return JSONResponse({"ok": True, "added": len(body.issue_numbers)})
-        except RuntimeError as exc:
-            logger.error("Failed to add items to crate #%d: %s", crate_number, exc)
-            return JSONResponse(
-                {"error": "Failed to add items to crate"}, status_code=500
-            )
-
-    @router.delete("/api/crates/{crate_number}/items")
-    async def remove_crate_items(
-        crate_number: int, body: CrateItemsRequest
-    ) -> JSONResponse:
-        """Remove issues from a milestone (crate) by clearing their milestone.
-
-        Only clears the milestone if the issue is currently assigned to the
-        specified crate (milestone), avoiding unintended removal from a
-        different milestone.
-        """
-        try:
-            current_issues = await pr_manager.list_milestone_issues(crate_number)
-            current_nums = {i.get("number") for i in current_issues}
-            removed = 0
-            for issue_number in body.issue_numbers:
-                if issue_number in current_nums:
-                    await pr_manager.set_issue_milestone(issue_number, None)
-                    removed += 1
-            return JSONResponse({"ok": True, "removed": removed})
-        except RuntimeError as exc:
-            logger.error("Failed to remove items from crate #%d: %s", crate_number, exc)
-            return JSONResponse(
-                {"error": "Failed to remove items from crate"}, status_code=500
-            )
-
-    @router.get("/api/crates/active")
-    async def get_active_crate() -> JSONResponse:
-        """Return the active crate number, title, and progress."""
-        orch = get_orchestrator()
-        active_number = state.get_active_crate_number()
-        result: dict[str, Any] = {
-            "crate_number": active_number,
-            "title": None,
-            "progress": 0,
-            "open_issues": 0,
-            "closed_issues": 0,
-            "total_issues": 0,
-        }
-        if active_number is not None and orch is not None:
-            try:
-                crates = await pr_manager.list_milestones(state="all")
-                active = next((c for c in crates if c.number == active_number), None)
-                if active:
-                    total = active.open_issues + active.closed_issues
-                    result["title"] = active.title
-                    result["open_issues"] = active.open_issues
-                    result["closed_issues"] = active.closed_issues
-                    result["total_issues"] = total
-                    result["progress"] = (
-                        round(active.closed_issues / total * 100) if total > 0 else 0
-                    )
-            except Exception:
-                logger.warning("Failed to enrich active crate details", exc_info=True)
-        return JSONResponse(result)
-
-    @router.post("/api/crates/active")
-    async def set_active_crate(body: dict[str, Any]) -> JSONResponse:
-        """Set the active crate. Body: ``{"crate_number": N}`` or ``{"crate_number": null}``."""
-        crate_number = body.get("crate_number")
-        if crate_number is not None and not isinstance(crate_number, int):
-            return JSONResponse(
-                {
-                    "status": "error",
-                    "detail": "crate_number must be an integer or null",
-                },
-                status_code=400,
-            )
-        orch = get_orchestrator()
-        if orch is None:
-            # Fallback: update state directly when orchestrator isn't running
-            state.set_active_crate_number(crate_number)
-            return JSONResponse({"status": "ok", "crate_number": crate_number})
-        if crate_number is not None:
-            await orch.crate_manager.activate_crate(crate_number)
-        else:
-            state.set_active_crate_number(None)
-        return JSONResponse({"status": "ok", "crate_number": crate_number})
-
-    @router.post("/api/crates/advance")
-    async def advance_crate() -> JSONResponse:
-        """Advance past the current active crate to the next open one.
-
-        Calls ``check_and_advance()`` which completes the active crate
-        and activates the next milestone with open issues.  If the
-        current crate still has open issues, it is force-cleared first
-        so the pipeline moves forward regardless.
-        """
-        orch = get_orchestrator()
-        cm = orch.crate_manager if orch is not None else None
-        if cm is None:
-            state.set_active_crate_number(None)
-            return JSONResponse({"status": "ok", "previous": None, "next": None})
-        previous = cm.active_crate_number
-        # Force-clear first so check_and_advance will see no active
-        # crate (if it still has open issues, check_and_advance would
-        # be a no-op otherwise).
-        state.set_active_crate_number(None)
-        # Now find the next open crate
-        try:
-            crates = await pr_manager.list_milestones(state="open")
-            candidates = sorted(
-                (c for c in crates if c.open_issues > 0 and c.number != previous),
-                key=lambda c: c.number,
-            )
-            if candidates:
-                await cm.activate_crate(candidates[0].number)
-                return JSONResponse(
-                    {
-                        "status": "ok",
-                        "previous": previous,
-                        "next": candidates[0].number,
-                    }
-                )
-        except Exception:
-            logger.warning("Failed to find next crate during advance", exc_info=True)
-        return JSONResponse({"status": "ok", "previous": previous, "next": None})
-
-    @router.get("/api/hitl")
-    async def get_hitl(
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Fetch issues/PRs labeled for human-in-the-loop (stuck on CI)."""
-        if not _is_pipeline_active(repo):
-            return JSONResponse([])
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        orch = _get_orch()
-        # Use cached data when available
-        if orch and isinstance(getattr(orch, "github_cache", None), GitHubDataCache):
-            items = orch.github_cache.get_hitl_items()
-        else:
-            hitl_labels = list(
-                dict.fromkeys([*_cfg.hitl_label, *_cfg.hitl_active_label])
-            )
-            manager = _pr_manager_for(_cfg, _bus)
-            items = await manager.list_hitl_items(hitl_labels)
-        enriched = []
-        for item in items:
-            data = dict(item) if isinstance(item, dict) else item.model_dump()
-            issue_num: int = int(
-                data.get("issue", 0) if isinstance(item, dict) else item.issue
-            )
-            if orch:
-                data["status"] = orch.get_hitl_status(issue_num)
-            cause = _state.get_hitl_cause(issue_num)
-            origin = _state.get_hitl_origin(issue_num)
-            if not cause and origin:
-                if origin in _cfg.review_label:
-                    cause = "Review escalation"
-                elif origin in _cfg.find_label:
-                    cause = "Triage escalation"
-                else:
-                    cause = "Escalation (reason not recorded)"
-            if cause:
-                data["cause"] = cause
-            # Flag items held for issue type review
-            if cause and (
-                "epic detected" in cause.lower()
-                or "bug report detected" in cause.lower()
-            ):
-                data["issueTypeReview"] = True
-            cached_summary = state.get_hitl_summary(issue_num)
-            data["llmSummary"] = cached_summary or ""
-            data["llmSummaryUpdatedAt"] = state.get_hitl_summary_updated_at(issue_num)
-            visual_ev = state.get_hitl_visual_evidence(issue_num)
-            if visual_ev:
-                data["visualEvidence"] = visual_ev.model_dump()
-            if (
-                not cached_summary
-                and config.transcript_summarization_enabled
-                and not config.dry_run
-                and bool(config.gh_token)
-                and _hitl_summary_retry_due(issue_num)
-            ):
-                asyncio.create_task(
-                    _warm_hitl_summary(issue_num, cause=cause or "", origin=origin)
-                )
-            enriched.append(data)
-
-        return JSONResponse(enriched)
-
-    @router.get("/api/hitl/{issue_number}/summary")
-    async def get_hitl_summary(issue_number: int) -> JSONResponse:
-        """Return cached HITL summary, generating one if missing."""
-        cached = state.get_hitl_summary(issue_number)
-        if cached:
-            return JSONResponse(
-                {
-                    "issue": issue_number,
-                    "summary": cached,
-                    "updated_at": state.get_hitl_summary_updated_at(issue_number),
-                    "cached": True,
-                }
-            )
-
-        cause = state.get_hitl_cause(issue_number) or ""
-        origin = state.get_hitl_origin(issue_number)
-        summary = await _compute_hitl_summary(issue_number, cause=cause, origin=origin)
-        if summary:
-            return JSONResponse(
-                {
-                    "issue": issue_number,
-                    "summary": summary,
-                    "updated_at": state.get_hitl_summary_updated_at(issue_number),
-                    "cached": False,
-                }
-            )
-        return JSONResponse(
-            {
-                "issue": issue_number,
-                "summary": "",
-                "updated_at": None,
-                "cached": False,
-            }
-        )
-
-    @router.post("/api/hitl/{issue_number}/correct")
-    async def hitl_correct(issue_number: int, body: dict[str, Any]) -> JSONResponse:
-        """Submit a correction for a HITL issue to guide retry."""
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"status": "no orchestrator"}, status_code=400)
-        correction = body.get("correction") or ""
-        if not correction.strip():
-            return JSONResponse(
-                {"status": "error", "detail": "Correction text must not be empty"},
-                status_code=400,
-            )
-        orch.submit_hitl_correction(issue_number, correction)
-
-        # Swap labels for immediate dashboard feedback
-        await pr_manager.swap_pipeline_labels(issue_number, config.hitl_active_label[0])
-
-        await event_bus.publish(
-            HydraFlowEvent(
-                type=EventType.HITL_UPDATE,
-                data=HITLUpdatePayload(
-                    issue=issue_number,
-                    status="processing",
-                    action="correct",
-                ),
-            )
-        )
-        return JSONResponse({"status": "ok"})
-
-    # ------------------------------------------------------------------
-    # HITL helpers
-    # ------------------------------------------------------------------
-
-    def _clear_hitl_state(
-        orch: HydraFlowOrchestrator | None,
-        issue_number: int,
-    ) -> None:
-        """Clear all HITL tracking state for an issue.
-
-        Consolidates the repeated skip + remove pattern used by every
-        HITL resolution endpoint.
-        """
-        if orch:
-            orch.skip_hitl_issue(issue_number)
-        state.remove_hitl_origin(issue_number)
-        state.remove_hitl_cause(issue_number)
-        state.remove_hitl_summary(issue_number)
-
-    async def _resolve_hitl_item(
-        issue_number: int,
-        orch: HydraFlowOrchestrator,
-        *,
-        action: str,
-        comment_heading: str,
-        comment_body: str,
-        outcome_type: IssueOutcomeType,
-        reason: str,
-    ) -> JSONResponse:
-        """Clear HITL state, record outcome, post comment, and publish event.
-
-        Shared implementation for hitl_skip, hitl_close, and
-        hitl_approve_process which all follow the same
-        cleanup → outcome → comment → event → response pattern.
-
-        ``orch`` must be the already-fetched, non-None orchestrator from the
-        caller so that state cleanup is always paired with the side effects the
-        caller already performed (label swaps, issue close, etc.).
-        """
-        _clear_hitl_state(orch, issue_number)
-        state.record_outcome(
-            issue_number,
-            outcome_type,
-            reason=reason,
-            phase="hitl",
-        )
-
-        try:
-            await pr_manager.post_comment(
-                issue_number,
-                f"**{comment_heading}** — {comment_body}\n\n---\n*HydraFlow Dashboard*",
-            )
-        except Exception:  # noqa: BLE001
-            logger.warning(
-                "Failed to post %s comment for issue #%d",
-                action,
-                issue_number,
-                exc_info=True,
-            )
-
-        await event_bus.publish(
-            HydraFlowEvent(
-                type=EventType.HITL_UPDATE,
-                data=HITLUpdatePayload(
-                    issue=issue_number,
-                    status="resolved",
-                    action=action,
-                    reason=reason,
-                ),
-            )
-        )
-        return JSONResponse({"status": "ok"})
-
-    # ------------------------------------------------------------------
-    # HITL route handlers
-    # ------------------------------------------------------------------
-
-    @router.post("/api/hitl/{issue_number}/skip")
-    async def hitl_skip(issue_number: int, body: HITLSkipRequest) -> JSONResponse:
-        """Remove a HITL issue from the queue without action (reason required)."""
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"status": "no orchestrator"}, status_code=400)
-
-        # Read origin before clearing state
-        state.get_hitl_origin(issue_number)
-
-        await pr_manager.close_issue(issue_number)
-
-        return await _resolve_hitl_item(
-            issue_number,
-            orch,
-            action="skip",
-            comment_heading="HITL Skip",
-            comment_body=f"Operator skipped this issue.\n\n**Reason:** {body.reason}",
-            outcome_type=IssueOutcomeType.HITL_SKIPPED,
-            reason=body.reason,
-        )
-
-    @router.post("/api/hitl/{issue_number}/close")
-    async def hitl_close(issue_number: int, body: HITLCloseRequest) -> JSONResponse:
-        """Close a HITL issue on GitHub (reason required)."""
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"status": "no orchestrator"}, status_code=400)
-        await pr_manager.close_issue(issue_number)
-
-        return await _resolve_hitl_item(
-            issue_number,
-            orch,
-            action="close",
-            comment_heading="HITL Close",
-            comment_body=f"Operator closed this issue.\n\n**Reason:** {body.reason}",
-            outcome_type=IssueOutcomeType.HITL_CLOSED,
-            reason=body.reason,
-        )
-
-    @router.post("/api/hitl/{issue_number}/approve-process")
-    async def hitl_approve_process(issue_number: int) -> JSONResponse:
-        """Approve a HITL item held for issue type review.
-
-        All issue types (bugs, epics, etc.) route to triage first.
-        """
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"status": "no orchestrator"}, status_code=400)
-
-        target_label = config.find_label[0]
-        target_stage = "triage"
-
-        await pr_manager.swap_pipeline_labels(issue_number, target_label)
-
-        return await _resolve_hitl_item(
-            issue_number,
-            orch,
-            action="approved_for_processing",
-            comment_heading="Approved for processing",
-            comment_body=(
-                f"Operator approved this issue.\n\n"
-                f"Routing to **{target_stage}** (`{target_label}`)."
-            ),
-            outcome_type=IssueOutcomeType.HITL_APPROVED,
-            reason=f"Operator approved issue type for processing ({target_stage})",
-        )
-
-    @router.get("/api/human-input")
-    async def get_human_input_requests(
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Return pending human-input prompts from the orchestrator."""
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        orch = _get_orch()
-        if orch:
-            return JSONResponse(orch.human_input_requests)
-        return JSONResponse({})
-
-    @router.post("/api/human-input/{issue_number}")
-    async def provide_human_input(
-        issue_number: int,
-        body: dict[str, Any],
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Submit an operator answer to a pending human-input request."""
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        orch = _get_orch()
-        if orch:
-            answer = body.get("answer", "")
-            orch.provide_human_input(issue_number, answer)
-            return JSONResponse({"status": "ok"})
-        return JSONResponse({"status": "no orchestrator"}, status_code=400)
-
-    @router.post("/api/control/start")
-    async def start_orchestrator() -> JSONResponse:
-        """Create and start a new orchestrator instance."""
-        orch = get_orchestrator()
-        if orch and orch.running:
-            return JSONResponse({"error": "already running"}, status_code=409)
-
-        from orchestrator import HydraFlowOrchestrator
-
-        # Remove pipeline workers from the disabled set — pipeline gating
-        # is handled solely by pipeline_enabled flag, not bg_worker_enabled.
-        existing_disabled = state.get_disabled_workers()
-        pipeline_names = set(_DEFAULT_PIPELINE_WORKERS)
-        cleaned = existing_disabled - pipeline_names
-        if cleaned != existing_disabled:
-            state.set_disabled_workers(cleaned)
-
-        new_orch = HydraFlowOrchestrator(
-            config,
-            event_bus=event_bus,
-            state=state,
-            pipeline_enabled=False,
-        )
-        set_orchestrator(new_orch)
-        set_run_task(asyncio.create_task(new_orch.run()))
-        await event_bus.publish(
-            HydraFlowEvent(
-                type=EventType.ORCHESTRATOR_STATUS,
-                data=OrchestratorStatusPayload(status="running", reset=True),
-            )
-        )
-        return JSONResponse({"status": "started"})
-
-    @router.post("/api/control/stop")
-    async def stop_orchestrator() -> JSONResponse:
-        """Request a graceful stop of the running orchestrator."""
-        orch = get_orchestrator()
-        if not orch or not orch.running:
-            return JSONResponse({"error": "not running"}, status_code=400)
-        await orch.request_stop()
-        return JSONResponse({"status": "stopping"})
-
-    @router.post("/api/control/clear-credit-pause")
-    async def clear_credit_pause() -> JSONResponse:
-        """Clear an active credit pause, waking any sleeping loops."""
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"error": "no orchestrator"}, status_code=400)
-        if orch.credits_paused_until is None:
-            return JSONResponse({"error": "not paused"}, status_code=400)
-        orch.clear_credit_pause()
-        return JSONResponse({"status": "cleared"})
-
-    @router.get("/api/control/status")
-    async def get_control_status(
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Return orchestrator run status, config summary, and version info."""
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        orch = _get_orch()
-        status = "idle"
-        current_session = None
-        latest_version = ""
-        update_available = False
-        if orch:
-            status = orch.run_status
-            current_session = orch.current_session_id
-        update_result = load_cached_update_result(current_version=get_app_version())
-        if update_result is not None:
-            latest_version = update_result.latest_version or ""
-            update_available = update_result.update_available
-        credits_until = (
-            orch.credits_paused_until.isoformat()
-            if orch and orch.credits_paused_until
-            else None
-        )
-        try:
-            control_status = ControlStatus(status)
-        except ValueError:
-            control_status = ControlStatus.IDLE
-        response = ControlStatusResponse(
-            status=control_status,
-            credits_paused_until=credits_until,
-            config=ControlStatusConfig(
-                app_version=get_app_version(),
-                latest_version=latest_version,
-                update_available=update_available,
-                repo=_cfg.repo,
-                ready_label=_cfg.ready_label,
-                find_label=_cfg.find_label,
-                planner_label=_cfg.planner_label,
-                review_label=_cfg.review_label,
-                hitl_label=_cfg.hitl_label,
-                hitl_active_label=_cfg.hitl_active_label,
-                fixed_label=_cfg.fixed_label,
-                max_triagers=_cfg.max_triagers,
-                max_workers=_cfg.max_workers,
-                max_planners=_cfg.max_planners,
-                max_reviewers=_cfg.max_reviewers,
-                max_hitl_workers=_cfg.max_hitl_workers,
-                batch_size=_cfg.batch_size,
-                model=_cfg.model,
-                pr_unstick_batch_size=_cfg.pr_unstick_batch_size,
-                worktree_base=str(_cfg.worktree_base),
-            ),
-        )
-        data = response.model_dump()
-        data["current_session_id"] = current_session
-        return JSONResponse(data)
-
-    @router.post("/api/control/credit-refresh")
-    async def credit_refresh(
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Attempt to clear credit pause and resume processing.
-
-        Probes the Anthropic API first.  If credits are still exhausted the
-        pause is kept and the client receives ``{"status": "still_exhausted"}``
-        so the UI can display immediate feedback.
-        """
-        from subprocess_util import probe_credit_availability
-
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        orch = _get_orch()
-        if not orch:
-            return JSONResponse({"error": "no orchestrator"}, status_code=400)
-        if orch.credits_paused_until is None:
-            return JSONResponse({"status": "not_paused"})
-        # Probe the API to see if credits are actually available now.
-        credits_available = await probe_credit_availability()
-        if not credits_available:
-            return JSONResponse({"status": "still_exhausted"})
-        cleared = orch.try_clear_credit_pause()
-        if not cleared:
-            return JSONResponse({"status": "not_paused"})
-        return JSONResponse({"status": "resuming"})
-
-    @router.post("/api/admin/prep")
-    async def admin_prep(
-        repo: str | None = Query(default=None, description="Repo slug to target"),
-    ) -> JSONResponse:
-        return await _execute_admin_task("prep", run_prep, repo)
-
-    @router.post("/api/admin/scaffold")
-    async def admin_scaffold(
-        repo: str | None = Query(default=None, description="Repo slug to target"),
-    ) -> JSONResponse:
-        return await _execute_admin_task("scaffold", run_scaffold, repo)
-
-    @router.post("/api/admin/clean")
-    async def admin_clean(
-        repo: str | None = Query(default=None, description="Repo slug to target"),
-    ) -> JSONResponse:
-        return await _execute_admin_task("clean", run_clean, repo)
-
-    @router.post("/api/admin/ensure-labels")
-    async def admin_ensure_labels(
-        repo: str | None = Query(default=None, description="Repo slug to target"),
-    ) -> JSONResponse:
-        return await _execute_admin_task("ensure-labels", run_ensure_labels, repo)
-
-    # Mutable fields that can be changed at runtime via PATCH
-    _MUTABLE_FIELDS = {
-        "max_triagers",
-        "max_workers",
-        "max_planners",
-        "max_reviewers",
-        "max_hitl_workers",
-        "model",
-        "review_model",
-        "planner_model",
-        "batch_size",
-        "max_ci_fix_attempts",
-        "max_quality_fix_attempts",
-        "max_review_fix_attempts",
-        "min_review_findings",
-        "max_merge_conflict_fix_attempts",
-        "ci_check_timeout",
-        "ci_poll_interval",
-        "poll_interval",
-        "pr_unstick_interval",
-        "pr_unstick_batch_size",
-        "unstick_auto_merge",
-        "unstick_all_causes",
-        "memory_auto_approve",
-        "worktree_base",
-    }
-
-    @router.patch("/api/control/config")
-    async def patch_config(
-        body: dict[str, Any],
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Update runtime config fields. Pass ``persist: true`` to save to disk.
-
-        When *repo* is provided, updates are scoped to that runtime's config and
-        persisted as overrides in the repo store.
-        """
-        persist = body.pop("persist", False)
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-
-        updates: dict[str, Any] = {}
-
-        for key, value in body.items():
-            if key not in _MUTABLE_FIELDS:
-                continue
-            if not hasattr(_cfg, key):
-                continue
-            updates[key] = value
-
-        if not updates:
-            return JSONResponse({"status": "ok", "updated": {}})
-
-        # Validate updates through Pydantic field constraints
-        test_values = _cfg.model_dump()
-        test_values.update(updates)
-        try:
-            validated = HydraFlowConfig.model_validate(test_values)
-        except ValidationError as exc:
-            errors = exc.errors()
-            msg = "; ".join(
-                f"{e['loc'][-1]}: {e['msg']}" for e in errors if e.get("loc")
-            )
-            return JSONResponse(
-                {"status": "error", "message": msg or "Invalid configuration"},
-                status_code=422,
-            )
-
-        # Apply validated values to the live config
-        applied: dict[str, Any] = {}
-        for key in updates:
-            validated_value = getattr(validated, key)
-            object.__setattr__(_cfg, key, validated_value)
-            applied[key] = validated_value
-
-        if applied:
-            if repo and repo_store is not None:
-                repo_store.update_overrides(repo, applied)
-            elif persist:
-                save_config_file(_cfg.config_file, applied)
-
-        return JSONResponse({"status": "ok", "updated": applied})
-
-    # Known workers with human-friendly labels (pipeline loops + background)
-    _bg_worker_defs = [
-        (
-            "triage",
-            "Triage",
-            "Classifies freshly discovered issues and routes them into the pipeline.",
-        ),
-        (
-            "plan",
-            "Plan",
-            "Builds implementation plans for triaged issues that are ready to execute.",
-        ),
-        (
-            "implement",
-            "Implement",
-            "Runs coding agents to implement planned issues and open pull requests.",
-        ),
-        (
-            "review",
-            "Review",
-            "Reviews PRs, applies fixes, and merges approved work when checks pass.",
-        ),
-        (
-            "memory_sync",
-            "Memory Manager",
-            "Ingests memory and transcript issues into durable learnings and proposals.",
-        ),
-        (
-            "retrospective",
-            "Retrospective",
-            "Captures post-merge outcomes and identifies recurring delivery patterns.",
-        ),
-        (
-            "review_insights",
-            "Review Insights",
-            "Aggregates recurring review feedback into improvement opportunities.",
-        ),
-        (
-            "pipeline_poller",
-            "Pipeline Poller",
-            "Refreshes live pipeline snapshots for dashboard queue/status rendering.",
-        ),
-        (
-            "pr_unsticker",
-            "PR Unsticker",
-            "Requeues stalled HITL PRs by validating requirements and reopening flow.",
-        ),
-        (
-            "report_issue",
-            "Report Issue",
-            "Processes queued bug reports into GitHub issues via the configured agent.",
-        ),
-        (
-            "adr_reviewer",
-            "ADR Reviewer",
-            "Reviews proposed ADRs via a 3-judge council and routes to accept, reject, or escalate.",
-        ),
-    ]
-
-    # Workers that have independent configurable intervals
-    _INTERVAL_WORKERS = {
-        "memory_sync",
-        "pr_unsticker",
-        "pipeline_poller",
-        "report_issue",
-    }
-    # Pipeline loops share poll_interval (read-only display)
-    _PIPELINE_WORKERS = {"triage", "plan", "implement", "review"}
-    _WORKER_SOURCE_ALIASES: dict[str, tuple[str, ...]] = {
-        "plan": ("planner",),
-        "implement": ("agent",),
-        "review": ("reviewer", "merge_conflict", "fresh_rebuild"),
-    }
-
-    def _build_system_worker_inference_stats() -> dict[str, dict[str, int]]:
-        """Aggregate prompt-telemetry inference stats keyed by worker name."""
-        telemetry = PromptTelemetry(config)
-        source_totals = telemetry.get_source_totals()
-
-        worker_totals: dict[str, dict[str, int]] = {}
-        for worker_name, _label, _description in _bg_worker_defs:
-            sources = (worker_name, *_WORKER_SOURCE_ALIASES.get(worker_name, ()))
-            totals = {
-                "inference_calls": 0,
-                "total_tokens": 0,
-                "pruned_chars_total": 0,
-            }
-            for source_name in sources:
-                source_entry = source_totals.get(source_name)
-                if not source_entry:
-                    continue
-                totals["inference_calls"] += source_entry["inference_calls"]
-                totals["total_tokens"] += source_entry["total_tokens"]
-                totals["pruned_chars_total"] += source_entry["pruned_chars_total"]
-            if totals["inference_calls"] > 0:
-                saved_tokens_est = round(totals["pruned_chars_total"] / 4)
-                worker_totals[worker_name] = {
-                    "inference_calls": totals["inference_calls"],
-                    "total_tokens": totals["total_tokens"],
-                    "pruned_chars_total": totals["pruned_chars_total"],
-                    "saved_tokens_est": saved_tokens_est,
-                    "unpruned_tokens_est": totals["total_tokens"] + saved_tokens_est,
-                }
-        return worker_totals
-
-    def _compute_next_run(
-        last_run: str | None, interval_seconds: int | None
-    ) -> str | None:
-        """Compute next run ISO timestamp from last_run + interval."""
-        if not last_run or not interval_seconds:
-            return None
-        try:
-            last_dt = datetime.fromisoformat(last_run)
-            if last_dt.tzinfo is None:
-                last_dt = last_dt.replace(tzinfo=UTC)
-            next_dt = last_dt + timedelta(seconds=interval_seconds)
-            return next_dt.isoformat()
-        except (ValueError, TypeError):
-            return None
-
-    @router.get("/api/system/workers")
-    async def get_system_workers(
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Return last known status of each background worker."""
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        orch = _get_orch()
-        bg_states = orch.get_bg_worker_states() if orch else {}
-        persisted_states: dict[str, BackgroundWorkerState] = {}
-        if not orch:
-            try:
-                persisted_states = _state.get_bg_worker_states()
-            except Exception:  # pragma: no cover - defensive guard
-                logger.exception("Failed to load persisted bg worker states")
-        inference_by_worker = _build_system_worker_inference_stats()
-        workers = []
-        for name, label, description in _bg_worker_defs:
-            enabled = orch.is_bg_worker_enabled(name) if orch else True
-
-            # Determine interval for this worker
-            interval: int | None = None
-            if name in _INTERVAL_WORKERS and orch:
-                interval = orch.get_bg_worker_interval(name)
-            elif name in _INTERVAL_WORKERS:
-                if name == "memory_sync":
-                    interval = _cfg.memory_sync_interval
-                elif name == "pr_unsticker":
-                    interval = _cfg.pr_unstick_interval
-                elif name == "pipeline_poller":
-                    interval = 5
-            elif name in _PIPELINE_WORKERS:
-                interval = _cfg.poll_interval
-
-            entry = bg_states.get(name) or persisted_states.get(name)
-            if entry:
-                last_run = entry.get("last_run")
-                raw_details = entry.get("details", {})
-                details: dict[str, Any] = (
-                    dict(raw_details)
-                    if isinstance(raw_details, dict)
-                    else {"raw_details": str(raw_details)}
-                )
-                details.update(inference_by_worker.get(name, {}))
-                workers.append(
-                    BackgroundWorkerStatus(
-                        name=name,
-                        label=label,
-                        description=description,
-                        status=BGWorkerHealth(
-                            entry.get("status", BGWorkerHealth.DISABLED)
-                        ),
-                        enabled=enabled,
-                        last_run=last_run,
-                        interval_seconds=interval,
-                        next_run=_compute_next_run(last_run, interval),
-                        details=details,
-                    )
-                )
-            else:
-                workers.append(
-                    BackgroundWorkerStatus(
-                        name=name,
-                        label=label,
-                        description=description,
-                        enabled=enabled,
-                        interval_seconds=interval,
-                        details=inference_by_worker.get(name, {}),
-                    )
-                )
-        return JSONResponse(BackgroundWorkersResponse(workers=workers).model_dump())
-
-    @router.post("/api/control/bg-worker")
-    async def toggle_bg_worker(body: dict[str, Any]) -> JSONResponse:
-        """Enable or disable a background worker."""
-        name = body.get("name")
-        enabled = body.get("enabled")
-        if not name or enabled is None:
-            return JSONResponse(
-                {"error": "name and enabled are required"}, status_code=400
-            )
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"error": "no orchestrator"}, status_code=400)
-        orch.set_bg_worker_enabled(name, bool(enabled))
-        return JSONResponse({"status": "ok", "name": name, "enabled": bool(enabled)})
-
-    @router.post("/api/control/bg-worker/trigger")
-    async def trigger_bg_worker(body: dict[str, Any]) -> JSONResponse:
-        """Trigger an immediate execution of a background worker."""
-        name = body.get("name")
-        if not name:
-            return JSONResponse({"error": "name is required"}, status_code=400)
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"error": "no orchestrator"}, status_code=400)
-        triggered = orch.trigger_bg_worker(name)
-        if not triggered:
-            return JSONResponse({"error": f"unknown worker '{name}'"}, status_code=404)
-        return JSONResponse({"status": "ok", "name": name})
-
-    @router.post("/api/control/bg-worker/interval")
-    async def set_bg_worker_interval(body: dict[str, Any]) -> JSONResponse:
-        """Update the polling interval for a background worker."""
-        name = body.get("name")
-        interval = body.get("interval_seconds")
-        if not name or interval is None:
-            return JSONResponse(
-                {"error": "name and interval_seconds are required"}, status_code=400
-            )
-        if name not in _INTERVAL_BOUNDS:
-            return JSONResponse(
-                {"error": f"interval not editable for worker '{name}'"}, status_code=400
-            )
-        try:
-            interval = int(interval)
-        except (TypeError, ValueError):
-            return JSONResponse(
-                {"error": "interval_seconds must be an integer"}, status_code=400
-            )
-        lo, hi = _INTERVAL_BOUNDS[name]
-        if interval < lo or interval > hi:
-            return JSONResponse(
-                {"error": f"interval_seconds must be between {lo} and {hi}"},
-                status_code=422,
-            )
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"error": "no orchestrator"}, status_code=400)
-        orch.set_bg_worker_interval(name, interval)
-        return JSONResponse(
-            {"status": "ok", "name": name, "interval_seconds": interval}
-        )
-
-    @router.get("/api/bot-pr/settings")
-    async def get_bot_pr_settings() -> JSONResponse:
-        """Return current bot PR auto-merge settings."""
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"error": "no orchestrator"}, status_code=400)
-        settings = orch.state.get_bot_pr_settings()
-        return JSONResponse(settings.model_dump())
-
-    @router.post("/api/bot-pr/settings")
-    async def set_bot_pr_settings(body: dict[str, Any]) -> JSONResponse:
-        """Update bot PR auto-merge settings."""
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"error": "no orchestrator"}, status_code=400)
-
-        current = orch.state.get_bot_pr_settings()
-        update = current.model_dump()
-        for key in ("authors", "failure_strategy", "review_mode"):
-            if key in body:
-                update[key] = body[key]
-
-        try:
-            from models import BotPRSettings  # noqa: PLC0415
-
-            new_settings = BotPRSettings(**update)
-        except (ValueError, ValidationError) as exc:
-            return JSONResponse({"error": str(exc)}, status_code=400)
-
-        orch.state.set_bot_pr_settings(new_settings)
-        return JSONResponse({"status": "ok", **new_settings.model_dump()})
-
-    @router.get("/api/issues/outcomes")
-    async def get_issue_outcomes() -> JSONResponse:
-        """Return all recorded issue outcomes."""
-        outcomes = state.get_all_outcomes()
-        return JSONResponse({k: v.model_dump() for k, v in outcomes.items()})
+    _register_metrics(router, ctx)
 
     # --- Issue history cache ---
     # Cache the aggregated issue_rows + pr_to_issue for the unfiltered case.
@@ -2870,276 +1753,6 @@ def create_router(
             ).model_dump()
         )
 
-    @router.get("/api/metrics")
-    async def get_metrics(
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Return lifetime stats, derived rates, time-to-merge, and thresholds."""
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        lifetime = _state.get_lifetime_stats()
-        rates: dict[str, float] = {}
-        total_reviews = (
-            lifetime.total_review_approvals + lifetime.total_review_request_changes
-        )
-        if lifetime.issues_completed > 0:
-            rates["merge_rate"] = lifetime.prs_merged / lifetime.issues_completed
-            rates["quality_fix_rate"] = (
-                lifetime.total_quality_fix_rounds / lifetime.issues_completed
-            )
-            rates["hitl_escalation_rate"] = (
-                lifetime.total_hitl_escalations / lifetime.issues_completed
-            )
-            rates["avg_implementation_seconds"] = (
-                lifetime.total_implementation_seconds / lifetime.issues_completed
-            )
-        if total_reviews > 0:
-            rates["first_pass_approval_rate"] = (
-                lifetime.total_review_approvals / total_reviews
-            )
-            rates["reviewer_fix_rate"] = lifetime.total_reviewer_fixes / total_reviews
-        time_to_merge = _state.get_merge_duration_stats()
-        thresholds = _state.check_thresholds(
-            _cfg.quality_fix_rate_threshold,
-            _cfg.approval_rate_threshold,
-            _cfg.hitl_rate_threshold,
-        )
-        retries = _state.get_retries_summary()
-        if retries:
-            rates["retries_per_stage"] = sum(retries.values())
-
-        telemetry = PromptTelemetry(_cfg)
-        inference_lifetime = telemetry.get_lifetime_totals()
-        orch = _get_orch()
-        session_id = orch.current_session_id if orch else ""
-        inference_session = (
-            telemetry.get_session_totals(session_id) if session_id else {}
-        )
-
-        return JSONResponse(
-            MetricsResponse(
-                lifetime=lifetime,
-                rates=rates,
-                time_to_merge=time_to_merge,
-                thresholds=thresholds,
-                inference_lifetime=inference_lifetime,
-                inference_session=inference_session,
-            ).model_dump()
-        )
-
-    @router.get("/api/metrics/github")
-    async def get_github_metrics(
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Query GitHub for issue/PR counts by label state."""
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        manager = _pr_manager_for(_cfg, _bus)
-        counts = await manager.get_label_counts(_cfg)
-        return JSONResponse(counts)
-
-    @router.get("/api/metrics/history")
-    async def get_metrics_history(
-        repo: RepoSlugParam = None,
-    ) -> JSONResponse:
-        """Historical snapshots from the metrics issue + current in-memory snapshot.
-
-        Falls back to local disk cache when the orchestrator is not running.
-        """
-        _cfg, _state, _bus, _get_orch = _resolve_runtime(repo)
-        orch = _get_orch()
-        if orch is None:
-            # Serve from local cache without requiring the orchestrator
-            snapshots = _load_local_metrics_cache(_cfg)
-            return JSONResponse(
-                MetricsHistoryResponse(snapshots=snapshots).model_dump()
-            )
-        mgr = orch.metrics_manager
-        snapshots = await mgr.fetch_history_from_issue()
-        current = mgr.latest_snapshot
-        return JSONResponse(
-            MetricsHistoryResponse(
-                snapshots=snapshots,
-                current=current,
-            ).model_dump()
-        )
-
-    @router.get("/api/runs")
-    async def list_run_issues() -> JSONResponse:
-        """Return issue numbers that have recorded runs."""
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse([])
-        return JSONResponse(orch.run_recorder.list_issues())
-
-    @router.get("/api/runs/{issue_number}")
-    async def get_runs(issue_number: int) -> JSONResponse:
-        """Return all recorded runs for an issue."""
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse([])
-        runs = orch.run_recorder.list_runs(issue_number)
-        return JSONResponse([r.model_dump() for r in runs])
-
-    @router.get("/api/runs/{issue_number}/{timestamp}/{filename}")
-    async def get_run_artifact(
-        issue_number: int, timestamp: str, filename: str
-    ) -> Response:
-        """Return a specific artifact file from a recorded run."""
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"error": "no orchestrator"}, status_code=400)
-        content = orch.run_recorder.get_run_artifact(issue_number, timestamp, filename)
-        if content is None:
-            return JSONResponse({"error": "artifact not found"}, status_code=404)
-        return Response(content=content, media_type="text/plain")
-
-    @router.get("/api/artifacts/stats")
-    async def get_artifact_stats() -> JSONResponse:
-        """Return storage statistics for run artifacts."""
-        orch = get_orchestrator()
-        if not orch:
-            return JSONResponse({"error": "no orchestrator"}, status_code=400)
-        stats = orch.run_recorder.get_storage_stats()
-        stats["retention_days"] = config.artifact_retention_days
-        stats["max_size_mb"] = config.artifact_max_size_mb
-        return JSONResponse(stats)
-
-    @router.get("/api/harness-insights")
-    async def get_harness_insights() -> JSONResponse:
-        """Return recent harness failure patterns and improvement suggestions."""
-        from harness_insights import (
-            HarnessInsightStore,
-            generate_suggestions,
-        )
-
-        memory_dir = config.data_path("memory")
-        store = HarnessInsightStore(memory_dir)
-        records = store.load_recent(config.harness_insight_window)
-        proposed = store.get_proposed_patterns()
-        suggestions = generate_suggestions(
-            records, config.harness_pattern_threshold, proposed
-        )
-
-        # Build category summary
-        cat_counts: Counter[str] = Counter(r.category for r in records)
-        sub_counts: Counter[str] = Counter()
-        for r in records:
-            for sub in r.subcategories:
-                sub_counts[sub] += 1
-
-        return JSONResponse(
-            {
-                "total_failures": len(records),
-                "category_counts": dict(cat_counts.most_common()),
-                "subcategory_counts": dict(sub_counts.most_common()),
-                "suggestions": [s.model_dump() for s in suggestions],
-                "proposed_patterns": sorted(proposed),
-            }
-        )
-
-    @router.get("/api/harness-insights/history")
-    async def get_harness_insights_history() -> JSONResponse:
-        """Return raw failure records for historical analysis."""
-        from harness_insights import HarnessInsightStore
-
-        memory_dir = config.data_path("memory")
-        store = HarnessInsightStore(memory_dir)
-        records = store.load_recent(config.harness_insight_window)
-        return JSONResponse([r.model_dump() for r in records])
-
-    @router.get("/api/review-insights")
-    async def get_review_insights() -> JSONResponse:
-        """Return aggregated review feedback patterns and category breakdown."""
-        from review_insights import ReviewInsightStore, analyze_patterns
-
-        memory_dir = config.data_path("memory")
-        store = ReviewInsightStore(memory_dir)
-        records = store.load_recent(config.review_insight_window)
-        proposed = store.get_proposed_categories()
-
-        verdict_counts: Counter[str] = Counter(r.verdict.value for r in records)
-        category_counts: Counter[str] = Counter(
-            cat for r in records for cat in r.categories
-        )
-        fixes_made_count = sum(1 for r in records if r.fixes_made)
-
-        patterns_raw = analyze_patterns(records, config.harness_pattern_threshold)
-        patterns = [
-            {
-                "category": cat,
-                "count": cnt,
-                "evidence": [
-                    {
-                        "issue_number": r.issue_number,
-                        "pr_number": r.pr_number,
-                        "summary": r.summary,
-                    }
-                    for r in evidence
-                ],
-            }
-            for cat, cnt, evidence in patterns_raw
-        ]
-
-        return JSONResponse(
-            {
-                "total_reviews": len(records),
-                "verdict_counts": dict(verdict_counts),
-                "category_counts": dict(category_counts),
-                "fixes_made_count": fixes_made_count,
-                "patterns": patterns,
-                "proposed_categories": sorted(proposed),
-            }
-        )
-
-    @router.get("/api/retrospectives")
-    async def get_retrospectives() -> JSONResponse:
-        """Return aggregated retrospective stats and recent entries."""
-        from retrospective import RetrospectiveEntry
-
-        retro_path = config.data_path("memory", "retrospectives.jsonl")
-        entries: list[RetrospectiveEntry] = []
-        if retro_path.exists():
-            for line in retro_path.read_text().strip().splitlines():
-                with contextlib.suppress(Exception):
-                    entries.append(RetrospectiveEntry.model_validate_json(line))
-        entries = entries[-config.retrospective_window :]
-
-        if not entries:
-            return JSONResponse(
-                {
-                    "total_entries": 0,
-                    "avg_plan_accuracy": 0,
-                    "avg_quality_fix_rounds": 0,
-                    "avg_ci_fix_rounds": 0,
-                    "avg_duration_seconds": 0,
-                    "reviewer_fix_rate": 0,
-                    "verdict_counts": {},
-                    "entries": [],
-                }
-            )
-
-        n = len(entries)
-        avg_accuracy = round(sum(e.plan_accuracy_pct for e in entries) / n, 1)
-        avg_quality = round(sum(e.quality_fix_rounds for e in entries) / n, 2)
-        avg_ci = round(sum(e.ci_fix_rounds for e in entries) / n, 2)
-        avg_duration = round(sum(e.duration_seconds for e in entries) / n, 1)
-        fix_count = sum(1 for e in entries if e.reviewer_fixes_made)
-        verdict_counts: Counter[str] = Counter(
-            str(e.review_verdict) for e in entries if e.review_verdict
-        )
-
-        return JSONResponse(
-            {
-                "total_entries": n,
-                "avg_plan_accuracy": avg_accuracy,
-                "avg_quality_fix_rounds": avg_quality,
-                "avg_ci_fix_rounds": avg_ci,
-                "avg_duration_seconds": avg_duration,
-                "reviewer_fix_rate": round(fix_count / n, 3),
-                "verdict_counts": dict(verdict_counts),
-                "entries": [e.model_dump() for e in entries],
-            }
-        )
-
     @router.get("/api/memories")
     async def get_memories() -> JSONResponse:
         """Return memory items from local JSONL event log."""
@@ -3205,615 +1818,10 @@ def create_router(
         timelines = state.get_all_completed_timelines()
         return JSONResponse([t.model_dump() for t in timelines.values()])
 
-    # --- Repo runtime lifecycle endpoints ---
+    # --- State/runtimes/repos/filesystem routes (extracted to _state_routes.py) ---
+    from dashboard_routes._state_routes import register as _register_state
 
-    def _is_default_repo(slug: str) -> bool:
-        """Check if *slug* refers to the default (host) repo."""
-        return ctx._is_default_repo(slug)
-
-    @router.get("/api/runtimes")
-    async def list_runtimes() -> JSONResponse:
-        """List all registered repo runtimes with status."""
-
-        infos: list[dict[str, Any]] = []
-
-        # Always include the default (host) repo.
-        orch = get_orchestrator()
-        pipeline_active = _default_repo_pipeline_running()
-        default_slug = config.repo.replace("/", "-") if config.repo else ""
-        if config.repo:
-            infos.append(
-                RepoRuntimeInfo(
-                    slug=default_slug,
-                    repo=config.repo,
-                    running=pipeline_active,
-                    session_id=orch.current_session_id
-                    if orch and orch.running
-                    else None,
-                ).model_dump()
-            )
-
-        if registry is not None:
-            for rt in registry.all:
-                infos.append(
-                    RepoRuntimeInfo(
-                        slug=rt.slug,
-                        repo=rt.config.repo,
-                        running=rt.running,
-                        session_id=rt.orchestrator.current_session_id
-                        if rt.running
-                        else None,
-                    ).model_dump()
-                )
-        return JSONResponse({"runtimes": infos})
-
-    @router.get("/api/runtimes/{slug}")
-    async def get_runtime_status(slug: str) -> JSONResponse:
-        """Get status of a specific repo runtime."""
-
-        if _is_default_repo(slug):
-            orch = get_orchestrator()
-            pipeline_active = _default_repo_pipeline_running()
-            default_slug = config.repo.replace("/", "-") if config.repo else ""
-            info = RepoRuntimeInfo(
-                slug=default_slug,
-                repo=config.repo,
-                running=pipeline_active,
-                session_id=orch.current_session_id if orch and orch.running else None,
-            )
-            return JSONResponse(info.model_dump())
-
-        if registry is None:
-            return JSONResponse(
-                {"error": "No runtime registry configured"}, status_code=501
-            )
-        rt = registry.get(slug)
-        if rt is None:
-            return JSONResponse({"error": f"Unknown repo: {slug}"}, status_code=404)
-        info = RepoRuntimeInfo(
-            slug=rt.slug,
-            repo=rt.config.repo,
-            running=rt.running,
-            session_id=rt.orchestrator.current_session_id if rt.running else None,
-        )
-        return JSONResponse(info.model_dump())
-
-    _DEFAULT_PIPELINE_WORKERS = ("triage", "plan", "implement", "review", "hitl")
-
-    def _default_repo_pipeline_running() -> bool:
-        """Return True if the default repo's pipeline is enabled."""
-        orch = get_orchestrator()
-        if not orch or not orch.running:
-            return False
-        return orch.pipeline_enabled
-
-    @router.post("/api/runtimes/{slug}/start")
-    async def start_runtime(slug: str) -> JSONResponse:
-        """Start a specific repo runtime."""
-        if _is_default_repo(slug):
-            orch = get_orchestrator()
-            if not orch or not orch.running:
-                return JSONResponse(
-                    {"error": "Orchestrator not running"}, status_code=400
-                )
-            orch.pipeline_enabled = True
-            return JSONResponse({"status": "started", "slug": slug})
-
-        if registry is None:
-            return JSONResponse(
-                {"error": "No runtime registry configured"}, status_code=501
-            )
-        rt = registry.get(slug)
-        if rt is None:
-            return JSONResponse({"error": f"Unknown repo: {slug}"}, status_code=404)
-        if rt.running:
-            return JSONResponse({"error": "Already running"}, status_code=409)
-        await rt.start()
-        return JSONResponse({"status": "started", "slug": slug})
-
-    @router.post("/api/runtimes/{slug}/stop")
-    async def stop_runtime(slug: str) -> JSONResponse:
-        """Stop a specific repo runtime."""
-        if _is_default_repo(slug):
-            orch = get_orchestrator()
-            if not orch or not orch.running:
-                return JSONResponse({"error": "Not running"}, status_code=400)
-            orch.pipeline_enabled = False
-            return JSONResponse({"status": "stopped", "slug": slug})
-
-        if registry is None:
-            return JSONResponse(
-                {"error": "No runtime registry configured"}, status_code=501
-            )
-        rt = registry.get(slug)
-        if rt is None:
-            return JSONResponse({"error": f"Unknown repo: {slug}"}, status_code=404)
-        if not rt.running:
-            return JSONResponse({"error": "Not running"}, status_code=400)
-        await rt.stop()
-        return JSONResponse({"status": "stopped", "slug": slug})
-
-    @router.delete("/api/runtimes/{slug}")
-    async def remove_runtime(slug: str) -> JSONResponse:
-        """Stop and unregister a repo runtime."""
-        if registry is None:
-            return JSONResponse(
-                {"error": "No runtime registry configured"}, status_code=501
-            )
-        rt = registry.get(slug)
-        if rt is None:
-            return JSONResponse({"error": f"Unknown repo: {slug}"}, status_code=404)
-        if remove_repo_cb is not None:
-            try:
-                await remove_repo_cb(slug)
-            except Exception as exc:  # noqa: BLE001
-                logger.warning("remove_repo callback failed for %s: %s", slug, exc)
-                return JSONResponse({"error": "Failed to remove repo"}, status_code=500)
-            return JSONResponse({"status": "removed", "slug": slug})
-        if rt.running:
-            await rt.stop()
-        registry.remove(slug)
-        if repo_store is not None:
-            try:
-                repo_store.remove(slug)
-            except Exception:  # noqa: BLE001
-                logger.warning(
-                    "Failed to remove repo %s from store", slug, exc_info=True
-                )
-        return JSONResponse({"status": "removed", "slug": slug})
-
-    # --- Multi-repo endpoints ---
-
-    @router.get("/api/repos")
-    async def list_supervised_repos() -> JSONResponse:
-        """List repos from the store or callback."""
-        if repo_store is not None or list_repos_cb is not None:
-            records = _list_repo_records()
-            payload: list[dict[str, Any]] = []
-            for rec in records:
-                runtime = registry.get(rec.slug) if registry else None
-                safe_slug = rec.slug.replace("/", "-") if rec.slug else rec.slug
-                payload.append(
-                    {
-                        "slug": safe_slug,
-                        "repo": rec.repo,
-                        "path": rec.path,
-                        "running": bool(runtime.running) if runtime else False,
-                        "session_id": runtime.orchestrator.current_session_id
-                        if runtime and runtime.running
-                        else None,
-                    }
-                )
-            return JSONResponse({"repos": payload, "can_register": True})
-        return JSONResponse({"repos": [], "can_register": False})
-
-    _root_names: dict[int, str] = {0: "Home", 1: "Temp"}
-
-    @router.get("/api/fs/roots")
-    async def list_browsable_roots() -> JSONResponse:
-        """Return filesystem roots that are safe to browse from the UI."""
-        all_roots = _repo_roots_fn()
-        roots = [
-            {"name": _root_names.get(i, f"Root {i + 1}"), "path": root}
-            for i, root in enumerate(all_roots)
-        ]
-        return JSONResponse({"roots": roots})
-
-    @router.get("/api/fs/list")
-    async def list_browsable_directories(
-        path: str | None = Query(default=None),
-    ) -> JSONResponse:
-        """List child directories for the requested path under allowed roots."""
-        allowed_roots = _repo_roots_fn()
-        if not allowed_roots:
-            return JSONResponse(
-                {"error": "no allowed roots configured"}, status_code=500
-            )
-        target_raw = path or allowed_roots[0]
-        target_path, error = _normalize_allowed_dir(
-            target_raw, allowed_roots=allowed_roots
-        )
-        if error or target_path is None:
-            return JSONResponse({"error": error or "invalid path"}, status_code=400)
-
-        current = str(target_path)
-        parent: str | None = None
-        parent_candidate = os.path.realpath(str(target_path.parent))
-        inside_allowed_parent = any(
-            parent_candidate == root or parent_candidate.startswith(f"{root}{os.sep}")
-            for root in allowed_roots
-        )
-        if inside_allowed_parent and parent_candidate != current:
-            parent = parent_candidate
-
-        directories: list[dict[str, str]] = []
-        try:
-            for child in sorted(target_path.iterdir(), key=lambda p: p.name.lower()):
-                if not child.is_dir():
-                    continue
-                # Hide dot-directories in the default browser view.
-                if child.name.startswith("."):
-                    continue
-                child_real = os.path.realpath(str(child))
-                inside_allowed_child = any(
-                    child_real == root or child_real.startswith(f"{root}{os.sep}")
-                    for root in allowed_roots
-                )
-                if not inside_allowed_child:
-                    continue
-                directories.append({"name": child.name, "path": child_real})
-        except OSError as exc:
-            logger.warning("Failed to list directory %s: %s", target_path, exc)
-            return JSONResponse({"error": "failed to list directory"}, status_code=500)
-
-        return JSONResponse(
-            {
-                "current_path": current,
-                "parent_path": parent,
-                "directories": directories,
-            }
-        )
-
-    @router.post("/api/repos")
-    async def ensure_repo() -> JSONResponse:
-        """Legacy endpoint — supervisor feature removed (issue #2205)."""
-        return JSONResponse({"error": "supervisor unavailable"}, status_code=503)
-
-    @router.delete("/api/repos/{slug}")
-    async def remove_repo(slug: str) -> JSONResponse:
-        """Remove a repo via the callback, or directly from the store."""
-        if remove_repo_cb is not None:
-            try:
-                removed = await remove_repo_cb(slug)
-            except Exception as exc:  # noqa: BLE001
-                logger.warning("remove_repo callback failed: %s", exc)
-                return JSONResponse({"error": "Failed to remove repo"}, status_code=500)
-            if not removed:
-                return JSONResponse({"error": "Repo not found"}, status_code=404)
-            return JSONResponse({"status": "ok"})
-        # Fallback: remove directly from the persistent store (no orchestrator needed)
-        if repo_store is not None:
-            if repo_store.remove(slug):
-                return JSONResponse({"status": "ok"})
-            return JSONResponse({"error": "Repo not found"}, status_code=404)
-        return JSONResponse({"error": "supervisor unavailable"}, status_code=503)
-
-    async def _detect_repo_slug_from_path(repo_path: Path) -> str | None:  # noqa: PLR0911
-        """Extract ``owner/repo`` from git remote origin URL at *repo_path*."""
-        from urllib.parse import urlparse  # noqa: PLC0415
-
-        try:
-            proc = await asyncio.create_subprocess_exec(
-                "git",
-                "-C",
-                str(repo_path),
-                "remote",
-                "get-url",
-                "origin",
-                stdout=asyncio.subprocess.PIPE,
-                stderr=asyncio.subprocess.PIPE,
-            )
-            stdout, _ = await asyncio.wait_for(proc.communicate(), timeout=10)
-        except (FileNotFoundError, OSError, TimeoutError):
-            return None
-        url = (stdout or b"").decode().strip()
-        if not url:
-            return None
-        if url.startswith(("http://", "https://")):
-            parsed = urlparse(url)
-            host = (parsed.hostname or "").lower()
-            if host != "github.com":
-                return None
-            return parsed.path.lstrip("/").removesuffix(".git") or None
-        if url.startswith("git@"):
-            if "@" not in url or ":" not in url:
-                return None
-            user_host, _, remainder = url.partition(":")
-            _, _, host = user_host.partition("@")
-            if host.lower() != "github.com":
-                return None
-            slug = remainder.lstrip("/").removesuffix(".git")
-            return slug or None
-        return None
-
-    @router.post("/api/repos/add")
-    async def add_repo_by_path(  # noqa: PLR0911
-        req: dict[str, Any] | None = Body(default=None),
-        req_query: str | None = Query(default=None, alias="req"),
-        path: str | None = Query(default=None),
-        repo_path_query: str | None = Query(default=None, alias="repo_path"),
-    ) -> JSONResponse:
-        """Register a repo by local filesystem path (does NOT start it)."""
-        if isinstance(req, dict):
-            for key in ("path", "repo_path"):
-                value = req.get(key)
-                if value is not None and not isinstance(value, str):
-                    return JSONResponse(
-                        {"error": "path must be a string"}, status_code=400
-                    )
-            nested = req.get("req")
-            if isinstance(nested, dict):
-                for key in ("path", "repo_path"):
-                    value = nested.get(key)
-                    if value is not None and not isinstance(value, str):
-                        return JSONResponse(
-                            {"error": "path must be a string"}, status_code=400
-                        )
-        raw_path = _extract_repo_path(req, req_query, path, repo_path_query)
-        if not raw_path:
-            return JSONResponse({"error": "path required"}, status_code=400)
-        repo_path, path_error = _normalize_allowed_dir(
-            raw_path, allowed_roots=_repo_roots_fn()
-        )
-        if path_error or repo_path is None:
-            return JSONResponse(
-                {"error": path_error or "invalid path"}, status_code=400
-            )
-        # Validate it's a git repo
-        is_git = False
-        try:
-            proc = await asyncio.create_subprocess_exec(
-                "git",
-                "-C",
-                str(repo_path),
-                "rev-parse",
-                "--git-dir",
-                stdout=asyncio.subprocess.PIPE,
-                stderr=asyncio.subprocess.PIPE,
-            )
-            await asyncio.wait_for(proc.communicate(), timeout=10)
-            is_git = proc.returncode == 0
-        except (FileNotFoundError, OSError, TimeoutError):
-            pass
-        if not is_git:
-            return JSONResponse(
-                {"error": f"not a git repository: {raw_path}"},
-                status_code=400,
-            )
-        # Detect slug
-        slug = await _detect_repo_slug_from_path(repo_path)
-        if register_repo_cb is not None:
-            try:
-                record, repo_cfg = await register_repo_cb(repo_path, slug)
-            except ValueError as exc:
-                return JSONResponse({"error": str(exc)}, status_code=400)
-            except Exception as exc:  # noqa: BLE001
-                logger.warning("register_repo callback failed: %s", exc)
-                return JSONResponse(
-                    {"error": "Failed to register repo"}, status_code=500
-                )
-            labels_created = False
-            if slug:
-                try:
-                    from prep import ensure_labels  # noqa: PLC0415
-
-                    await ensure_labels(repo_cfg)
-                    labels_created = True
-                except Exception:  # noqa: BLE001
-                    logger.warning("Label creation failed for %s", slug, exc_info=True)
-            return JSONResponse(
-                {
-                    "status": "ok",
-                    "slug": record.slug,
-                    "path": record.path,
-                    "labels_created": labels_created,
-                }
-            )
-
-        # Supervisor fallback removed (issue #2205) — no register_repo_cb means 503.
-        return JSONResponse(
-            {"error": "supervisor unavailable"},
-            status_code=503,
-        )
-
-    @router.post("/api/repos/pick-folder")
-    async def pick_repo_folder() -> JSONResponse:
-        """Open a native folder picker and return the selected path."""
-        selected = await _pick_folder_with_dialog()
-        if not selected:
-            return JSONResponse({"error": "No folder selected"}, status_code=400)
-        path = Path(os.path.realpath(os.path.expanduser(selected)))
-        if not path.is_dir():
-            return JSONResponse(
-                {"error": "Selected path is not a directory"}, status_code=400
-            )
-        return JSONResponse({"path": str(path)})
-
-    @router.get("/api/github/repos")
-    async def list_github_repos(
-        query: str | None = Query(default=None),
-    ) -> JSONResponse:
-        """List GitHub repos for the authenticated user via ``gh repo list``."""
-        cmd = [
-            "gh",
-            "repo",
-            "list",
-            "--json",
-            "name,owner,url,description",
-            "--limit",
-            "100",
-        ]
-        proc: asyncio.subprocess.Process | None = None
-        try:
-            proc = await asyncio.create_subprocess_exec(
-                *cmd,
-                stdout=asyncio.subprocess.PIPE,
-                stderr=asyncio.subprocess.PIPE,
-            )
-            stdout, stderr = await asyncio.wait_for(proc.communicate(), timeout=30)
-        except FileNotFoundError:
-            return JSONResponse(
-                {
-                    "error": "gh CLI not found — install GitHub CLI and run 'gh auth login'"
-                },
-                status_code=503,
-            )
-        except TimeoutError:
-            if proc is not None:
-                with contextlib.suppress(ProcessLookupError):
-                    proc.kill()
-            return JSONResponse(
-                {"error": "gh CLI timed out"},
-                status_code=504,
-            )
-        if proc.returncode != 0:
-            msg = (stderr or b"").decode().strip()
-            if "auth" in msg.lower() or "login" in msg.lower():
-                return JSONResponse(
-                    {"error": "Not authenticated — run 'gh auth login' first"},
-                    status_code=401,
-                )
-            return JSONResponse(
-                {"error": f"gh repo list failed: {msg}"},
-                status_code=502,
-            )
-        try:
-            repos = json.loads(stdout or b"[]")
-        except json.JSONDecodeError:
-            return JSONResponse(
-                {"error": "Failed to parse gh output"},
-                status_code=502,
-            )
-        # Filter by query if provided
-        if query:
-            q = query.lower()
-            repos = [
-                r
-                for r in repos
-                if q in (r.get("name") or "").lower()
-                or q in ((r.get("owner") or {}).get("login") or "").lower()
-                or q
-                in f"{(r.get('owner') or {}).get('login', '')}/{r.get('name', '')}".lower()
-            ]
-        return JSONResponse({"repos": repos})
-
-    @router.post("/api/github/clone")
-    async def clone_github_repo(  # noqa: PLR0911
-        req: dict[str, Any] | None = Body(default=None),
-    ) -> JSONResponse:
-        """Clone a GitHub repo into the workspace directory and register it."""
-        if not isinstance(req, dict):
-            return JSONResponse({"error": "request body required"}, status_code=400)
-        slug = (req.get("slug") or "").strip()
-        if not slug or "/" not in slug:
-            return JSONResponse(
-                {"error": "slug required in owner/repo format"},
-                status_code=400,
-            )
-        raw_owner, raw_repo = slug.split("/", 1)
-        if not raw_owner or not raw_repo:
-            return JSONResponse(
-                {"error": "slug required in owner/repo format"},
-                status_code=400,
-            )
-        # Validate path components to prevent directory traversal
-        if (
-            not _SAFE_SLUG_COMPONENT.match(raw_owner)
-            or not _SAFE_SLUG_COMPONENT.match(raw_repo)
-            or set(raw_owner) <= {"."}
-            or set(raw_repo) <= {"."}
-        ):
-            return JSONResponse(
-                {"error": "slug contains invalid characters"},
-                status_code=400,
-            )
-        # Sanitise: extract only the final path component to break any
-        # traversal sequences.  PurePosixPath.name is recognised by
-        # CodeQL as a path-injection sanitiser.
-        owner = PurePosixPath(raw_owner).name
-        repo_name = PurePosixPath(raw_repo).name
-        if not owner or not repo_name:
-            return JSONResponse(
-                {"error": "slug contains invalid characters"},
-                status_code=400,
-            )
-        workspace_dir = Path(
-            os.path.expanduser(str(config.repos_workspace_dir))
-        ).resolve()
-        clone_target = (workspace_dir / owner / repo_name).resolve()
-        if not clone_target.is_relative_to(workspace_dir):
-            return JSONResponse(
-                {"error": "slug contains invalid characters"},
-                status_code=400,
-            )
-        # Reconstruct slug from sanitised components
-        slug = f"{owner}/{repo_name}"
-        already_cloned = clone_target.is_dir() and (clone_target / ".git").is_dir()
-        if not already_cloned:
-            workspace_dir.mkdir(parents=True, exist_ok=True)
-            owner_dir = clone_target.parent
-            owner_dir.mkdir(parents=True, exist_ok=True)
-            cmd = ["gh", "repo", "clone", slug, str(clone_target)]
-            clone_proc: asyncio.subprocess.Process | None = None
-            try:
-                clone_proc = await asyncio.create_subprocess_exec(
-                    *cmd,
-                    stdout=asyncio.subprocess.PIPE,
-                    stderr=asyncio.subprocess.PIPE,
-                )
-                _, stderr = await asyncio.wait_for(
-                    clone_proc.communicate(), timeout=300
-                )
-            except FileNotFoundError:
-                return JSONResponse(
-                    {"error": "gh CLI not found"},
-                    status_code=503,
-                )
-            except TimeoutError:
-                if clone_proc is not None:
-                    with contextlib.suppress(ProcessLookupError):
-                        clone_proc.kill()
-                return JSONResponse(
-                    {"error": "Clone timed out"},
-                    status_code=504,
-                )
-            if clone_proc.returncode != 0:
-                msg = (stderr or b"").decode().strip()
-                return JSONResponse(
-                    {"error": f"Clone failed: {msg}"},
-                    status_code=502,
-                )
-        # Register with the callback
-        if register_repo_cb is not None:
-            try:
-                record, repo_cfg = await register_repo_cb(clone_target, slug)
-            except ValueError as exc:
-                logger.warning("register_repo validation error: %s", exc)
-                return JSONResponse(
-                    {"error": "Invalid repository configuration"}, status_code=400
-                )
-            except Exception as exc:  # noqa: BLE001
-                logger.warning("register_repo callback failed: %s", exc)
-                return JSONResponse(
-                    {"error": "Failed to register repo"}, status_code=500
-                )
-            labels_created = False
-            try:
-                from prep import ensure_labels  # noqa: PLC0415
-
-                await ensure_labels(repo_cfg)
-                labels_created = True
-            except Exception:  # noqa: BLE001
-                logger.warning("Label creation failed for %s", slug, exc_info=True)
-            return JSONResponse(
-                {
-                    "status": "ok",
-                    "slug": record.slug,
-                    "path": record.path,
-                    "already_cloned": already_cloned,
-                    "labels_created": labels_created,
-                }
-            )
-        # Fallback: register via add_repo_by_path logic
-        return JSONResponse(
-            {
-                "status": "ok",
-                "slug": slug,
-                "path": str(clone_target),
-                "already_cloned": already_cloned,
-                "labels_created": False,
-            }
-        )
+    _register_state(router, ctx)
 
     @router.post("/api/intent")
     async def submit_intent(request: IntentRequest) -> JSONResponse:
@@ -3833,177 +1841,10 @@ def create_router(
         response = IntentResponse(issue_number=issue_number, title=title, url=url)
         return JSONResponse(response.model_dump())
 
-    @router.post("/api/report")
-    async def submit_report(request: ReportIssueRequest) -> JSONResponse:
-        """Queue a bug report for async processing by the report issue worker."""
-        report = PendingReport(
-            description=request.description,
-            screenshot_base64=request.screenshot_base64,
-            environment=request.environment,
-            reporter_id=request.reporter_id,
-        )
-        state.enqueue_report(report)
+    # --- Reports routes (extracted to _reports_routes.py) ---
+    from dashboard_routes._reports_routes import register as _register_reports
 
-        # Create a tracked report for the reporter if a reporter_id is provided
-        if request.reporter_id:
-            tracked = TrackedReport(
-                id=report.id,
-                reporter_id=request.reporter_id,
-                description=request.description,
-                status="queued",
-                history=[
-                    ReportHistoryEntry(
-                        action="submitted",
-                        detail="Bug report submitted via dashboard",
-                    )
-                ],
-            )
-            state.add_tracked_report(tracked)
-
-        # Trigger the report-issue worker immediately so the report
-        # doesn't wait for the next polling interval.
-        orch = get_orchestrator()
-        if orch is not None:
-            orch.trigger_bg_worker("report_issue")
-
-        title = f"[Bug Report] {request.description[:100]}"
-        response = ReportIssueResponse(
-            issue_number=0, title=title, url="", status="queued"
-        )
-        return JSONResponse(response.model_dump())
-
-    @router.get("/api/reports")
-    async def list_tracked_reports(
-        reporter_id: str = "", status: str | None = None
-    ) -> JSONResponse:
-        """List tracked reports for a given reporter, optionally filtered by status."""
-        if not reporter_id:
-            return JSONResponse([])
-        reports = state.get_tracked_reports(reporter_id, status=status)
-        return JSONResponse([r.model_dump() for r in reports])
-
-    @router.patch("/api/reports/{report_id}")
-    async def update_tracked_report(
-        report_id: str, body: TrackedReportUpdate
-    ) -> JSONResponse:
-        """Update a tracked report (confirm fixed, reopen, cancel)."""
-        report = state.get_tracked_report(report_id)
-        if report is None:
-            return JSONResponse({"error": "Report not found"}, status_code=404)
-        if (
-            body.reporter_id
-            and report.reporter_id
-            and body.reporter_id != report.reporter_id
-        ):
-            return JSONResponse({"error": "Forbidden"}, status_code=403)
-        # Validate state-machine transitions
-        valid_actions: dict[str, list[str]] = {
-            "confirm_fixed": ["fixed"],
-            "reopen": ["filed", "fixed", "in-progress", "closed"],
-            "cancel": ["queued", "in-progress", "filed", "fixed", "reopened"],
-        }
-        if report.status not in valid_actions.get(body.action, []):
-            return JSONResponse(
-                {
-                    "error": f"Action '{body.action}' is not allowed in status '{report.status}'"
-                },
-                status_code=422,
-            )
-        action_map: dict[
-            str,
-            tuple[
-                Literal[
-                    "queued",
-                    "in-progress",
-                    "filed",
-                    "fixed",
-                    "closed",
-                    "reopened",
-                ],
-                str,
-            ],
-        ] = {
-            "confirm_fixed": ("closed", "Confirmed fixed by reporter"),
-            "reopen": ("reopened", "Reopened by reporter"),
-            "cancel": ("closed", "Cancelled by reporter"),
-        }
-        status, default_detail = action_map[body.action]
-        updated = state.update_tracked_report(
-            report_id,
-            status=status,
-            detail=body.detail or default_detail,
-            action_label=body.action,
-        )
-        if updated is None:
-            return JSONResponse({"error": "Report not found"}, status_code=404)
-        return JSONResponse(updated.model_dump())
-
-    @router.get("/api/reports/{report_id}/history")
-    async def get_report_history(report_id: str) -> JSONResponse:
-        """Get the timeline/history for a tracked report."""
-        report = state.get_tracked_report(report_id)
-        if report is None:
-            return JSONResponse({"error": "Report not found"}, status_code=404)
-        return JSONResponse([entry.model_dump() for entry in report.history])
-
-    @router.post("/api/reports/refresh")
-    async def refresh_report_statuses(reporter_id: str = "") -> JSONResponse:
-        """Refresh statuses for filed and stale-queued reports.
-
-        * **Filed** reports: checks the linked GitHub issue state; if
-          the issue is closed the report transitions to ``"fixed"``.
-        * **Stale queued** reports (>30 min): re-enqueues them into the
-          pending queue so the background worker retries processing.
-        """
-        refreshed: list[dict[str, str]] = []
-
-        # --- filed → fixed when linked issue is closed as completed ---
-        for report in state.get_filed_reports():
-            if reporter_id and report.reporter_id != reporter_id:
-                continue
-            issue_number = _extract_issue_number(report.linked_issue_url)
-            if issue_number <= 0:
-                continue
-            issue_state = await pr_manager.get_issue_state(issue_number)
-            if issue_state == "COMPLETED":
-                state.update_tracked_report(
-                    report.id,
-                    status="fixed",
-                    action_label="fixed",
-                    detail=f"Issue #{issue_number} resolved",
-                )
-                refreshed.append({"id": report.id, "new_status": "fixed"})
-            elif issue_state == "NOT_PLANNED":
-                state.update_tracked_report(
-                    report.id,
-                    status="closed",
-                    action_label="closed",
-                    detail=f"Issue #{issue_number} closed as won't fix",
-                )
-                refreshed.append({"id": report.id, "new_status": "closed"})
-
-        # --- stale queued → re-enqueue pending ---
-        pending_ids = {p.id for p in state.get_pending_reports()}
-        for report in state.get_stale_queued_reports(stale_minutes=30):
-            if reporter_id and report.reporter_id != reporter_id:
-                continue
-            # Only re-enqueue if there's no pending entry already
-            if report.id not in pending_ids:
-                state.enqueue_report(
-                    PendingReport(
-                        id=report.id,
-                        description=report.description,
-                        reporter_id=report.reporter_id,
-                    )
-                )
-                state.update_tracked_report(
-                    report.id,
-                    action_label="retry",
-                    detail="Stale queued report re-enqueued for processing",
-                )
-                refreshed.append({"id": report.id, "new_status": "queued"})
-
-        return JSONResponse({"refreshed": refreshed})
+    _register_reports(router, ctx)
 
     @router.get("/api/sessions")
     async def get_sessions(
