+++ tests/test_dashboard_routes_repo.py
@@ -631,7 +631,7 @@ class TestPickRepoFolder:
         endpoint = find_endpoint(router, "/api/repos/pick-folder", "POST")
 
         with patch(
-            "dashboard_routes._routes._pick_folder_with_dialog",
+            "dashboard_routes._state_routes._pick_folder_with_dialog",
             new_callable=AsyncMock,
             return_value=None,
         ):
@@ -657,7 +657,7 @@ class TestPickRepoFolder:
         endpoint = find_endpoint(router, "/api/repos/pick-folder", "POST")
 
         with patch(
-            "dashboard_routes._routes._pick_folder_with_dialog",
+            "dashboard_routes._state_routes._pick_folder_with_dialog",
             new_callable=AsyncMock,
             return_value=str(repo_dir),
         ):
